日報、日付カラム追加

+++ db/migrate/20230430120318_create_individuals.rb
@@ -2,11 +2,10 @@ class CreateIndividuals < ActiveRecord::Migration[6.1]
   def change
     create_table :individuals do |t|
 
-      # t.integer :animal_species_id, null: false
-      t.string :name,               null: false
-      t.date :birthday,             null: false
-      t.integer :age
-      t.integer :weight
+      t.string  :name,              null: false
+      t.date    :birthday,          null: false
+      t.decimal :age
+      t.decimal :weight
 
       t.timestamps
     end
